requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-telegram-bot==21.5
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.0git add requirements.txt
git commit -m "add requirements"
git push